Lint scripts, rename variables, re-use getBaseRequestHeaders
