+++ package.json
@@ -12,8 +12,8 @@
     "build:prod": "NODE_ENV=production npm run build",
     "package": "npm run build:prod && npx @adobe/reactor-packager",
     "dev": "./scripts/dev.js",
-    "lint": "eslint \"*.js\" \"src/**/*.{js,jsx}\" \"test/**/*.{js,jsx}\"",
-    "format": "prettier --write \"*.{js,jsx}\" \"{src,test}/**/*.{js,jsx}\"",
+    "lint": "eslint \"*.js\" \"src/**/*.{js,jsx}\" \"test/**/*.{js,jsx}\" \"scripts/**/*.js\"",
+    "format": "prettier --write \"*.{js,jsx}\" \"{src,test}/**/*.{js,jsx}\" \"scripts/**/*.js\"",
     "test:unit": "karma start --single-run",
     "test:unit:coverage": "karma start --single-run --reporters spec,coverage",
     "test:unit:watch": "karma start",
