+++ scripts/token.js
@@ -1,10 +1,25 @@
+#!/usr/bin/env node
+/*
+Copyright 2020 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
 const auth = require("@adobe/jwt-auth");
 const adobeIOClientCredentials = require("../test/functional/helpers/adobeIOClientCredentials");
 
 (async () => {
   if (adobeIOClientCredentials) {
+    // eslint-disable-next-line no-console
     console.log(adobeIOClientCredentials.orgId);
-    result = await auth(adobeIOClientCredentials);
+    const result = await auth(adobeIOClientCredentials);
+    // eslint-disable-next-line no-console
     console.log(result.access_token);
   }
 })();
