+++ src/view/utils/getBaseRequestHeaders.js
