+++ test/docs/README.md
@@ -29,7 +29,7 @@ Within the Init tab on the right side of the page, find the `extensionSettings` 
 
 Some views may require authentication as well. When the extension runs inside Launch, Launch provides the extension the user's access token and org ID, which can then be used to access data from endpoints that require authentication. Because the local sandbox does not automatically provide the extension view a legitimate access token, you will need to obtain one and provide it to the extension view.
 
-One way of obtaining and using an access token is as follows: 
+#### OPTION 1 - Use the token from the Launch UI
 
 1. Log into Launch.
 1. Ensure you're on the org you'd like to be testing with.
@@ -42,11 +42,23 @@ One way of obtaining and using an access token is as follows:
 1. Paste the org ID as the value for the `orgId` property.
 1. Click the "Init" button. You may also need to refresh your browser before the view begins working correctly.
 
+#### OPTION 2 - Generate the token locally
+
+1. See the section below on Automated Testing to setup the two required environment variables
+1. `npm run token`
+1. Copy the second to last line of the token output. This is the IMS org. (It should end in @AdobeOrg)
+1. Go to the local testing sandbox.
+1. Paste the org ID as the value for the `orgId` property.
+1. Go back to the token command output. Copy the last line.
+1. Go back to the local testing sandbox.
+1. Paste the token as the `imsAccess` property.
+1. Click the "Init" button. You may also need to refresh your browser before the view begins working correctly.
+
 ## Automated Testing
 
 Automated testing runs some end-to-end tests that require authentication to the Adobe Experience Platform. Because we can't safely give public access to the Adobe Experience Platform account, these particular tests are restricted to Adobe employees who have been provided a private key to the Adobe I/O integration. When running tests, you may see a warning in your console indicating that some tests will be skipped because you don't have credentials set up. If you are not an Adobe employee and are working on related code, don't worry about running or updating these tests. Submit your pull request anyway and an Adobe employee will help ensure the code is properly tested.
 
-For Adobe employees, please reach out to the `alloy-engineering` Slack channel and someone will help you access the private key and client secret for the Adobe I/O integration. Save the private key somewhere on your hard drive. Then, create two environment variables: 
+For Adobe employees, please reach out to the `alloy-engineering` Slack channel and someone will help you access the private key and client secret for the Adobe I/O integration. Save the private key somewhere on your hard drive. Then, create two environment variables:
 
 * `EDGE_E2E_PRIVATE_KEY_FILE` - The value of this variable should be the path to the private key.
 * `EDGE_E2E_CLIENT_SECRET` - The value of this variable should be the client secret.
